masque: common items styling
problem with missing common items is solved.

as we not style them with nice grayish color.

but is it what we wanted to be ?
this kinda look good, but its not, we don't really want to see the white items border so highlited, we could prolly change alpha or change the border color of white items to be less distracting us from colored items...