masque: common items styling success
common items are now styled just like we wanted them to be.

problems remaining:
dim junk coloring of the itembutton main texture aka its icon is no more passed.
we can clearly see that the item icon of the junk items is no longer dimmed.

problem 2: junk items borders are too much distracting, i could say their color is kinda what we had before for white items, its light gray.

it is too distracting for the user. must be same color of the border and itemicon itself as it was in original adibags.